# app/api/openapi.yaml
...
openapi: 3.0.3
info:
  title: OpenWatch (Stage 0)
  version: 0.1.0-alpha.1
  description: |
    Stage 0 walking-skeleton surface — three endpoints that exercise the
    foundation stack (correlation, idempotency, audit, error envelope).
servers:
  - url: https://localhost:8443
    description: Local development
paths:
  /api/v1/auth/login:
    post:
      operationId: postAuthLogin
      summary: Username/password login with optional TOTP
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AuthLoginRequest'}
      responses:
        '200':
          description: Login succeeded; tokens issued
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AuthLoginResponse'}
        '401':
          description: Invalid credentials or MFA required
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
...
  /api/v1/auth/refresh:
    post:
      operationId: postAuthRefresh
      summary: Rotate the refresh token; mint a new access+refresh pair
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AuthRefreshRequest'}
      responses:
        '200':
          description: New token pair issued
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AuthLoginResponse'}
        '401':
          description: Refresh token invalid, expired, or reused (cascade revoke triggered)
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
  /api/v1/auth/me:
    get:
      operationId: getAuthMe
      summary: Return the calling identity
      responses:
        '200':
          description: Identity
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AuthMeResponse'}
        '401':
          description: No valid session or bearer
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
...
  /api/v1/auth/mfa:verify:
    post:
      operationId: postAuthMFAVerify
      summary: Confirm an enrolled MFA secret by verifying a generated OTP
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AuthMFAVerifyRequest'}
      responses:
        '204':
          description: MFA verified
        '401':
          description: MFA OTP invalid or replayed
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
  /api/v1/auth/password:change:
    post:
      operationId: postAuthPasswordChange
      summary: Change the calling user's password
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AuthPasswordChangeRequest'}
      responses:
        '204':
          description: Password updated
        '401':
          description: Current password wrong
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
        '400':
          description: New password rejected by policy
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
...
components:
...
  schemas:
    AuthLoginRequest:
      type: object
      required: [username, password]
      properties:
        username: {type: string, minLength: 1, maxLength: 256}
        password: {type: string, minLength: 1, maxLength: 256}
        otp:      {type: string, minLength: 6, maxLength: 6}
    AuthLoginResponse:
      type: object
      required: [access_token, refresh_token, user]
      properties:
        access_token:  {type: string}
        refresh_token: {type: string}
        user:          {$ref: '#/components/schemas/AuthMeResponse'}
    AuthRefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token: {type: string}
    AuthMeResponse:
      type: object
      required: [id, username, email, role, is_admin]
      properties:
        id:       {type: string, format: uuid}
        username: {type: string}
        email:    {type: string}
        role:     {type: string}
        is_admin: {type: boolean}
...
    AuthMFAVerifyRequest:
      type: object
      required: [otp]
      properties:
        otp: {type: string, minLength: 6, maxLength: 6}
    AuthPasswordChangeRequest:
      type: object
      required: [current_password, new_password]
      properties:
        current_password: {type: string, minLength: 1}
        new_password:     {type: string, minLength: 1}
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, fault, retryable, human_message]
          properties:
            code: {type: string}
            fault: {type: string, enum: [client, server, policy, external]}
            retryable: {type: boolean}
            human_message: {type: string}
            detail: {type: object, additionalProperties: true}
            correlation_id: {type: string}
...